#207 resolve core update issue missing labels (#211)
* move the git submodule to the latest

- v1.0.0 releease of loafwallet-core

# Conflicts:
#	Modules/loafwallet-core

* ISSUE: Not able to type and display recv addresses

- There is a failiure in the  conversion from CChar to Int8
- Have send working
- updated loafwallet-core v1.1.0
- Added falsePositived interface
- Added enum and set values for
 - low privacy
 - semi-private
 - anonymous
- Updated the loafwallet-core
- adds the fpRate var

* Resolved PeerManager set with proper optional setting.

* WIP: Cannot unwrap ltc1 addresses to retain them in the UILabel

- ltc1 addresses are present but not unwrapping properly
@Iferencak this is the disappearing error you saw when we updated the loafwallet-core.
- Not able to figure this out yet.
- All ideas are welcome

* WIP:  Still debugging

- The issue is the BRCore methods compare the `ltc1` address against the set of addresses and it fails because the BRSet is not setup for the address.

The issue is this worked before and I need to investigate what happened before and how things changed


* move the git submodule to the latest

- v1.0.0 releease of loafwallet-core

* Bugfix!: Duplicated char in converter

Added optic enum


* Move all the extensions from BRCore to separate files

- Moved these into separate files to help debug


* Refactored the conversion

- Fixed a bug where a element is duplicated
`unichar(charArray.20),unichar(charArray.21), unichar(charArray.22), unichar(charArray.23), unichar(charArray.23), unichar(charArray.24)`


---------
