+++ loafwallet/BRAPIClient+Wallet.swift
@@ -16,7 +16,6 @@ extension BRAPIClient {
 	func exchangeRates(isFallback: Bool = false, _ handler: @escaping (_ rates: [Rate], _ error: String?) -> Void)
 	{
 		let request = isFallback ? URLRequest(url: URL(string: fallbackRatesURL)!) : URLRequest(url: URL(string: APIServer.baseUrl + "v1/rates")!)
-		print("::: request: \(request.debugDescription)")
 		_ = dataTaskWithRequest(request) { data, _, error in
 			if error == nil, let data = data,
 			   let parsedData = try? JSONSerialization.jsonObject(with: data, options: .allowFragments)
