+++ loafwallet/TransactionCellViewModel.swift
@@ -75,7 +75,8 @@ class TransactionCellViewModel: ObservableObject {
 
 		feeText = transaction.amountDetails(isLtcSwapped: isLtcSwapped, rate: rate, rates: [rate], maxDigits: maxDigits)
 
-		addressText = String(format: transaction.direction.addressTextFormat, transaction.toAddress ?? "")
+		addressText = String(format: transaction.direction.addressTextFormat, transaction.toAddress ?? "DEBUG-UPDATED-ADDRESS")
+		print("::: toAddress: \(transaction.toAddress) raw : \(transaction.rawTransaction)")
 
 		if transaction.direction == .sent {
 			directionImageText = "arrowtriangle.up.circle.fill"
