+++ loafwallet/src/ViewModels/Transaction.swift
@@ -157,20 +157,36 @@ class Transaction {
 	lazy var toAddress: String? = {
 		switch self.direction {
 		case .sent:
-			guard let output = self.tx.outputs.filter({ output in
-				!self.wallet.containsAddress(output.swiftAddress)
-			}).first else { return nil }
-			return output.swiftAddress
+
+			guard let output = self
+				.tx.outputs.filter({ output in
+					!self.wallet.containsAddress(output.updatedSwiftAddress)
+				})
+				.first
+
+			else {
+				LWAnalytics.logEventWithParameters(itemName: ._20200112_ERR)
+				return nil
+			}
+
+			return output.updatedSwiftAddress
+
 		case .received:
 			guard let output = self.tx.outputs.filter({ output in
-				self.wallet.containsAddress(output.swiftAddress)
-			}).first else { return nil }
-			return output.swiftAddress
+				self.wallet.containsAddress(output.updatedSwiftAddress)
+
+			}).first
+			else {
+				LWAnalytics.logEventWithParameters(itemName: ._20200112_ERR)
+				return nil
+			}
+			return output.updatedSwiftAddress
+
 		case .moved:
 			guard let output = self.tx.outputs.filter({ output in
-				self.wallet.containsAddress(output.swiftAddress)
+				self.wallet.containsAddress(output.updatedSwiftAddress)
 			}).first else { return nil }
-			return output.swiftAddress
+			return output.updatedSwiftAddress
 		}
 	}()
